+++ code/classes/patricia.hpp
