+++ code/main.cpp
@@ -1,4 +1,5 @@
-#include "game.hpp"
+#include "classes/game.hpp"
+#include "files/import.hpp"
 #include <iostream>
 #include <fstream>
 #include <cstdio>
@@ -14,7 +15,7 @@ int main()
     std::getline(myfile, line);
 
     auto strings = cpd::customSplit(line, ';');
-    
+
     cpd::Game test(strings[0], strings[1], strings[4], strings[5], strings[3], strings[6], strings[2], strings[10]);
     std::cout << test << std::endl;
     myfile.close();
